admin: factor out raft routes
